Moved configuration to shared folder.
